+++ mmcls/models/losses/utils.py
@@ -32,7 +32,7 @@ def weight_reduce_loss(loss, weight=None, reduction='mean', avg_factor=None):
         loss (Tensor): Element-wise loss.
         weight (Tensor): Element-wise weights.
         reduction (str): Same as built-in losses of PyTorch.
-        avg_factor (float): Avarage factor when computing the mean of losses.
+        avg_factor (float): Average factor when computing the mean of losses.
 
     Returns:
         Tensor: Processed loss values.
