+++ mmcls/models/utils/inverted_residual.py
@@ -14,10 +14,10 @@ class InvertedResidual(BaseModule):
         in_channels (int): The input channels of this Module.
         out_channels (int): The output channels of this Module.
         mid_channels (int): The input channels of the depthwise convolution.
-        kernel_size (int): The kernal size of the depthwise convolution.
+        kernel_size (int): The kernel size of the depthwise convolution.
             Default: 3.
         stride (int): The stride of the depthwise convolution. Default: 1.
-        se_cfg (dict): Config dict for se layer. Defaul: None, which means no
+        se_cfg (dict): Config dict for se layer. Default: None, which means no
             se layer.
         conv_cfg (dict): Config dict for convolution layer. Default: None,
             which means using conv2d.
